[DP-207] fix: 불필요한 join 삭제 #216

+++ src/main/java/com/dapanda/chat/dto/response/ReadJoiningChatRoomResponse.java
@@ -16,7 +16,7 @@ public class ReadJoiningChatRoomResponse {
 	private LocalDateTime createdAt;
 	private LocalDateTime lastMessageAt;
 	private String lastMessage;
-	private Integer unreadCount;
+	private Long unreadCount;
 
 	//Member
 	private Long senderId;
